some enhancements
