+++ src/ws/io/directory.h
@@ -1,5 +1,13 @@
 #pragma once
 
+#ifdef _WIN32
+#include <windows.h>
+#else
+#include <dirent.h>
+#include <sys/stat.h>
+#include <unistd.h>
+#endif
+
 #include <algorithm>
 #include <string>
 #include <vector>
